pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.9",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.3.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.9",
    "wikipedia>=1.4.0",
    "pillow>=11.3.0",
]
